Update the code architecture according to the level.
