+++ exercises/hangman/.meta/src/reference/java/Hangman.java
@@ -44,8 +44,8 @@ class Hangman {
             Status.PLAYING);
     }
 
-    private static Hangman.Output processNewLetter(
-        final Hangman.Output state,
+    private static Output processNewLetter(
+        final Output state,
         final String letter) {
         if (state.guess.contains(letter) || state.misses.contains(letter)) {
             throw new IllegalArgumentException("Letter " + letter + " was already played");
@@ -81,58 +81,6 @@ class Hangman {
                 newParts,
                 newStatus);
         }
-    }
-
-    static class Output {
-
-        public final String secret;
-        public final String discovered;
-        public final Set<String> guess;
-        public final Set<String> misses;
-        public final List<Part> parts;
-        public final Status status;
-
-        Output(
-            final String secret,
-            final String discovered,
-            final Set<String> guess,
-            final Set<String> misses,
-            final List<Part> parts,
-            final Status status) {
-            this.secret = secret;
-            this.discovered = discovered;
-            this.guess = guess;
-            this.misses = misses;
-            this.parts = parts;
-            this.status = status;
-        }
-
-        @Override
-        public String toString() {
-            return "Output{" +
-                "secret='" + secret + '\'' +
-                ", discovered='" + discovered + '\'' +
-                ", guess=" + guess +
-                ", misses=" + misses +
-                ", parts=" + parts +
-                ", status=" + status +
-                '}';
-        }
-    }
-
-    enum Status {
-        PLAYING,
-        WIN,
-        LOSS
-    }
-
-    enum Part {
-        HEAD,
-        BODY,
-        LEFT_ARM,
-        RIGHT_ARM,
-        LEFT_LEG,
-        RIGHT_LEG
     }
 
     static final Part[] ORDER = Part.values();
