+++ exercises/hangman/src/main/java/Status.java
@@ -1,0 +1,12 @@
+/*
+ * (C) ActiveViam 2019
+ * ALL RIGHTS RESERVED. This material is the CONFIDENTIAL and PROPRIETARY
+ * property of ActiveViam. Any unauthorized use
+ * reproduction or transfer of this material is strictly prohibited
+ */
+
+enum Status {
+    PLAYING,
+    WIN,
+    LOSS
+}
